# static/openapi.json
openapi: 3.0.3
info:
  title: Betpreneur API
  version: 1.0.0
  description: "\n## Betpreneur - Sports Betting Intelligence Engine\n\nA Django REST
    API backend for the GrindAlgo betting intelligence system.\n\n### Authentication\nThis
    API uses JWT (JSON Web Tokens) for authentication.\n1. Obtain tokens via `/api/auth/login/`
    or `/api/auth/signup/`\n2. Include access token in header: `Authorization: Bearer
    <access_token>`\n3. Refresh expired tokens via `/api/auth/token/refresh/`\n\n###
    Key Features\n- **User Authentication**: Signup, login, email verification, password
    reset\n- **Algo Management**: Run betting algorithms, track picks\n- **Bankroll
    Tracking**: Monitor betting bankroll\n- **Reports**: Generate and track reports\n
    \   "
paths:
...
  /api/auth/reset-password/:
    post:
      operationId: auth_reset_password_create
      description: "\n        Reset password using token from email.\n        \n        **Requirements:**\n
        \       - Valid token (from forgot password email)\n        - New password
        + confirmation\n        \n        **Token expires:** 10 minutes\n        "
      summary: Reset password
      tags:
      - Authentication
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPassword'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/ResetPassword'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ResetPassword'
        required: true
      security:
      - jwtAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                description: Password reset successful
          description: ''
        '400':
          content:
            application/json:
              schema:
                description: Invalid or expired token
          description: ''
...
  /api/auth/verify-email/:
    post:
      operationId: auth_verify_email_create
      description: "\n        Verify user's email address with 6-digit code.\n        \n
        \       **Code sent to:** User's email address\n        **Expires in:** 10
        minutes\n        "
      summary: Verify email address
      tags:
      - Authentication
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyEmail'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/VerifyEmail'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/VerifyEmail'
        required: true
      security:
      - jwtAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                description: Email verified successfully
          description: ''
        '400':
          content:
            application/json:
              schema:
                description: Invalid or expired code
          description: ''
...
components:
  schemas:
...
    ResetPassword:
      type: object
      properties:
        token:
          type: string
          maxLength: 6
        user_id:
          type: string
          format: uuid
        new_password:
          type: string
          writeOnly: true
        confirm_password:
          type: string
          writeOnly: true
      required:
      - confirm_password
      - new_password
      - token
      - user_id
...
    VerifyEmail:
      type: object
      properties:
        code:
          type: string
          maxLength: 6
      required:
      - code
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT